# table-service/internal/adapter/http/response.go
package http

type createResponse struct {
	ID string `json:"id" example:"1923564209627467776"`
}

type createSessionResponse struct {
	URL string `json:"url" example:"http://localhost:3000?s=tYlC7uRGdaIT0-wDvCngv9qEWwM4Tdg4Jzoywe00fK2WhDqDHKCzqtybFDVgCPIGv1_isM3aXdb16KNGD4E-q2kewBRaVXZ5N7vgdi46Tc5po6_ZCHpFAT-Ei3xkKT0dL_f3Ruoiz9IzsnBhxlexXTEhKN9myrECimbKaDI="`
}